requirements.txt
gevent-socketio==0.3.6
gevent-websocket==0.9.3
gevent==1.0